feat: refactor review-related classes to use domain user entity and update response structure #5
