+++ src/main/java/com/babjo/deliverycommerce/domain/review/controller/ReviewController.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 import com.babjo.deliverycommerce.domain.review.dto.*;
 import com.babjo.deliverycommerce.domain.review.service.ReviewService;
-import com.babjo.deliverycommerce.global.common.dto.ApiResponse;
+import com.babjo.deliverycommerce.global.common.dto.CommonResponse;
 import com.babjo.deliverycommerce.global.security.UserPrincipal;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -23,41 +23,41 @@ public class ReviewController {
 
     @PostMapping
     @PreAuthorize("hasRole('CUSTOMER')")
-    public ResponseEntity<ApiResponse<ReviewCreateResponse>> createReview(
+    public ResponseEntity<CommonResponse<ReviewCreateResponse>> createReview(
             @AuthenticationPrincipal UserPrincipal principal,
             @Valid @RequestBody ReviewCreateRequest request
     ) {
-        return ApiResponse.created("리뷰 생성 성공", reviewService.createReview(principal, request));
+        return CommonResponse.created("리뷰 생성 성공", reviewService.createReview(principal, request));
     }
 
     @PutMapping("/{reviewId}")
     @PreAuthorize("hasAnyRole('CUSTOMER', 'MANAGER', 'MASTER')")
-    public ResponseEntity<ApiResponse<ReviewUpdateResponse>> updateReview(
+    public ResponseEntity<CommonResponse<ReviewUpdateResponse>> updateReview(
             @AuthenticationPrincipal UserPrincipal principal,
             @PathVariable UUID reviewId,
             @Valid @RequestBody ReviewUpdateRequest request
     ) {
-        return ApiResponse.ok("리뷰 수정 성공", reviewService.updateReview(principal, reviewId, request));
+        return CommonResponse.ok("리뷰 수정 성공", reviewService.updateReview(principal, reviewId, request));
     }
 
     @GetMapping
     @PreAuthorize("hasAnyRole('CUSTOMER', 'MANAGER', 'MASTER')")
-    public ResponseEntity<ApiResponse<List<ReviewResponse>>> getReviews(
+    public ResponseEntity<CommonResponse<List<ReviewResponse>>> getReviews(
             @AuthenticationPrincipal UserPrincipal principal,
             @RequestParam(required = false) UUID reviewId,  // 우선순위 1: 단건 조회
             @RequestParam(required = false) UUID storeId    // 우선순위 2: 가게별 목록 조회
     ) {
         // 파라미터 우선순위: reviewId > storeId > 전체(MANAGER/MASTER) 또는 본인(CUSTOMER/OWNER)
-        return ApiResponse.ok("리뷰 조회 성공", reviewService.getReviews(principal, reviewId, storeId));
+        return CommonResponse.ok("리뷰 조회 성공", reviewService.getReviews(principal, reviewId, storeId));
     }
 
     @DeleteMapping("/{reviewId}")
     @PreAuthorize("hasAnyRole('CUSTOMER', 'MANAGER', 'MASTER')")
-    public ResponseEntity<ApiResponse<Void>> deleteReview(
+    public ResponseEntity<CommonResponse<Void>> deleteReview(
             @AuthenticationPrincipal UserPrincipal principal,
             @PathVariable UUID reviewId
     ) {
         reviewService.deleteReview(reviewId, principal);
-        return ApiResponse.ok("리뷰 삭제 성공", null);
+        return CommonResponse.ok("리뷰 삭제 성공", null);
     }
 }
